+++ src/Http/Controllers/FormElementController.php
@@ -25,7 +25,7 @@ class FormElementController extends Controller
             $item = $model->getRepository()->find($id);
             if (is_null($item) || ! $model->isEditable($item)) {
                 return new JsonResponse([
-                    'message' => trans('lang.message.access_denied'),
+                    'message' => trans('sleeping_owl::lang.message.access_denied'),
                 ], 403);
             }
 
@@ -34,7 +34,7 @@ class FormElementController extends Controller
 
         if (! $model->isCreatable()) {
             return new JsonResponse([
-                'message' => trans('lang.message.access_denied'),
+                'message' => trans('sleeping_owl::lang.message.access_denied'),
             ], 403);
         }
 
@@ -53,7 +53,7 @@ class FormElementController extends Controller
             $item = $model->getRepository()->find($id);
             if (is_null($item) || ! $model->isEditable($item)) {
                 return new JsonResponse([
-                    'message' => trans('lang.message.access_denied'),
+                    'message' => trans('sleeping_owl::lang.message.access_denied'),
                 ], 403);
             }
 
@@ -62,7 +62,7 @@ class FormElementController extends Controller
 
         if (! $model->isCreatable()) {
             return new JsonResponse([
-                'message' => trans('lang.message.access_denied'),
+                'message' => trans('sleeping_owl::lang.message.access_denied'),
             ], 403);
         }