+++ resources/assets/js_owl/admin/form/selectajax.js
@@ -4,6 +4,14 @@ var isEmptyVal = function (value, trim) {
 
 Admin.Modules.register('form.elements.selectajax', () => {
     $('.js-data-ajax').each((e, item) => {
+
+        if (jQuery().select2) {
+            $.extend($.fn.select2.defaults.defaults.language, {
+                noResults: function() {return trans('lang.table.infoEmpty')},
+                inputTooShort: function() {return trans('lang.select.short', {min: $self.data('min-symbols')})}
+            });
+        }
+
         let options = {},
             $self = $(item);